chore: moved the routes out of the components folder
