+++ IndoorMapsFrontend/src/routes/BuildingEditor/FloorListItem.tsx
