+++ IndoorMapsFrontend/src/routes/BuildingEditor/GeomanControl.tsx
