+++ IndoorMapsFrontend/src/routes/BuildingEditor/styles/BuildingEditorBody.css
