+++ IndoorMapsFrontend/src/routes/BuildingViewer/Navigation/AreaNavigate.tsx
@@ -1,5 +1,5 @@
 import { Button, Switch, rem } from "@mantine/core";
-import { AreaToAreaRouteInfo } from "../../../../utils/types";
+import { AreaToAreaRouteInfo } from "../../../utils/types";
 import { IconArrowLeft, IconCurrentLocation, IconLocationShare } from "@tabler/icons-react";
 import AreaSearchBox from "./AreaSearchBox";
 import { AreaSearchBoxQuery$data } from "./__generated__/AreaSearchBoxQuery.graphql";
@@ -7,7 +7,7 @@ import { useEffect, useRef, useState } from "react";
 import { fetchQuery, graphql, useRelayEnvironment } from "react-relay";
 import { LatLng } from "leaflet";
 import { AreaNavigateAllDataQuery, AreaNavigateAllDataQuery$variables } from "./__generated__/AreaNavigateAllDataQuery.graphql";
-import { useUserLocation } from "../../../../utils/hooks";
+import { useUserLocation } from "../../../utils/hooks";
 
 const iconCurrentLocation = <IconCurrentLocation style={{ width: rem(16), height: rem(16) }} />
 const iconLocationShare = <IconLocationShare style={{ width: rem(16), height: rem(16) }} />
