+++ IndoorMapsFrontend/src/routes/BuildingViewer/ShareLocationModal.tsx
@@ -2,9 +2,9 @@ import { Button, Modal, TextInput, Group } from "@mantine/core";
 import { hasLength, useForm } from "@mantine/form";
 import { useEffect, useState } from "react";
 import { useParams } from "react-router-dom";
-import FormErrorNotification from "../../forms/FormErrorNotification";
-import { getPointBetweentwoPoints } from "../../../utils/utils";
-import { useUserLocation } from "../../../utils/hooks";
+import FormErrorNotification from "../../components/forms/FormErrorNotification";
+import { getPointBetweentwoPoints } from "../../utils/utils";
+import { useUserLocation } from "../../utils/hooks";
 import Cookies from 'js-cookie'
 
 const WS_ENDPOINT = import.meta.env.VITE_BACKEND_WEBSOCKET_URL;
