+++ IndoorMapsFrontend/src/routes/BuildingViewer/styles/BuildingViewer.css
