+++ IndoorMapsFrontend/src/routes/Directory/BuildingItem.tsx
