+++ IndoorMapsFrontend/src/routes/Directory/ConnectedBuildingItem.tsx
