+++ IndoorMapsFrontend/src/routes/Directory/styles/Directory.css
